Mudando a estrutura das classes
