+++ package.json
@@ -15,7 +15,8 @@
     "axios": "^0.21.4",
     "consign": "^0.1.6",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "mongoose": "^6.0.5"
   },
   "devDependencies": {
     "nodemon": "^2.0.12",
