+++ src/core/util/check-status.js
